MCP_Server/requirements.txt
mcp>=0.1.0
httpx
python-dotenv
pydantic
requests
cryptography
